+++ src/test/java/object/card/CardSymbolTest.java
@@ -1,6 +1,5 @@
-package object;
+package object.card;
 
-import object.card.CardSymbol;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
