+++ src/test/java/object/game/ScoreTypeTest.java
@@ -1,6 +1,5 @@
-package object;
+package object.game;
 
-import object.game.ScoreType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
