+++ src/test/java/object/participant/PlayerTest.java
@@ -1,6 +1,5 @@
-package object;
+package object.participant;
 
-import object.participant.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
